requirements.txt
absl-py
altair
altgraph
apache-beam
aplus
appdirs
appnope
APScheduler
argon2-cffi
argon2-cffi-bindings
asteval
astor
asttokens
astunparse
attrs
audioread
Automat
autosub
avro-python3
backcall
beautifulsoup4
beep
beepy
bleach
blinker
boto
boto3
botocore
bqplot
branca
bs4
cachetools
catboost
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
commonmark
constantly
contextlib2
crcmod
cryptography
cssselect
cvzone
cycler
Cython
dask
deap
debugpy
decorator
defusedxml
dill
distlib
dm-tree
docopt
docutils
entrypoints
executing
fastavro
ffmpeg-python
filelock
Flask
flatbuffers
fonttools
frozendict
future
fuzzywuzzy
gast
gin-config
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-pasta
googleapis-common-protos
graphviz
grpcio
gTTS
gunicorn
h5py
hdfs
httplib2
hyperband-snakemake
hyperlink
hyperopt
idna
imageai
imbalanced-learn
imblearn
importlib-metadata
importlib-resources
imutils
incremental
ipydatawidgets
ipykernel
ipython
ipython-genutils
ipywidgets
itemadapter
itemloaders
itsdangerous
jedi
Jinja2
jmespath
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-contrib-core
jupyter-contrib-nbextensions
jupyter-core
jupyter-highlight-selected-word
jupyter-latex-envs
jupyter-nbextensions-configurator
jupyterlab-pygments
jupyterlab-widgets
kaggle
keras
Keras-Preprocessing
keyring
kiwisolver
labelImg
libclang
librosa
LinearRegression
llvmlite
lmfit
lvis
lxml
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mediapipe
mistune
monty
msgpack-python
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
numba
numpy
oauth2client
oauthlib
opencv-contrib-python
opencv-python
opt-einsum
orjson
packaging
pandas
pandocfilters
parsel
parso
pexpect
pickleshare
Pillow
pkginfo
plotly
pooch
portalocker
progressbar
prometheus-client
promise
prompt-toolkit
Protego
proto-plus
protobuf
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycocotools
pycparser
pydash
pydeck
PyDispatcher
pydot
pydub
Pygments
pymongo
Pympler
pyOpenSSL
pyparsing
PyQt5
PyQt5-Qt5
PyQt5-sip
pyrsistent
pysrt
pytesseract
python-dateutil
python-Levenshtein
python-slugify
python-telegram-bot
pytz
pytz-deprecation-shim
PyYAML
pyzmq
qtconsole
QtPy
queuelib
readme-renderer
regex
requests
requests-oauthlib
requests-toolbelt
resampy
rfc3986
rich
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
sacrebleu
scikit-learn
scipy
Scrapy
selenium
semver
Send2Trash
sentencepiece
seqeval
service-identity
simpleaudio
six
sklearn
smmap
SoundFile
soupsieve
SpeechRecognition
stack-data
streamlit
tabulate
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-addons
tensorflow-datasets
tensorflow-hub
tensorflow-io
tensorflow-io-gcs-filesystem
tensorflow-metadata
tensorflow-model-optimization
tensorflow-object-detection-api
tensorflow-text
tensorflowjs
termcolor
terminado
testpath
text-unidecode
tf-estimator-nightly
tf-models-official
tf-slim
threadpoolctl
tk
toml
toolz
torch
torchvision
tornado
tqdm
traitlets
traittypes
twine
Twisted
typeguard
typing_extensions
tzdata
tzlocal
uncertainties
uritemplate
urllib3
validators
videocr
w3lib
watchtower
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
xmltodict
youtube-dl
zipp
zope.interface
